Move `bnk` module to library and refactor dynamic loading
This commit primarily refactors the `bnk` module from being part of the core to it's own separate dynamic library under `libs/bnk`, as well as changes its linkage to shared. It also modifies runtime to improve efficiency and facilitates the loading of dynamic libraries by refining the process in `import.cpp`. Additionally, it makes updates to `object.cpp` and `runtime.h` to enhance code readability and optimizes the usage of string data types in `lsp/main.cpp`. The commit cleans up redundant code and includes further support for native function documentation in `builtins.cpp` and fixes type casting in `object.h`.

+++ src/object.h
@@ -463,7 +463,7 @@ namespace bond {
 
         void patch_code(uint32_t offset, uint32_t oprand) { m_instructions[offset] = oprand; }
 
-        uint32_t current_index() { return m_instructions.size(); }
+        uint32_t current_index() { return (uint32_t)m_instructions.size(); }
 
         [[nodiscard]] t_string disassemble() const;
 
@@ -471,7 +471,7 @@ namespace bond {
 
         uint32_t get_code(size_t index) { return m_instructions[index]; }
 
-        uint32_t get_code_size() { return m_instructions.size(); }
+        uint32_t get_code_size() { return (uint32_t)m_instructions.size(); }
 
         SharedSpan get_span(size_t index) { return m_spans[index]; }
